add kover for code coverage (#63)
* add kover for code coverage

This commit introduces kover for code coverage reporting. It includes the necessary dependencies and configuration for generating coverage reports during builds. Additionally, it integrates with GitHub Actions to publish coverage reports on pull requests, enabling tracking and enforcement of code coverage thresholds.

* reduce coverage threshold

This commit reduces the minimum code coverage thresholds for line, branch, and instruction coverage to 20%. It also updates the GitHub workflow to reflect these changes.

* Updated CI workflow to run on macOS

This commit changes the GitHub Actions workflow to run on `macos-latest` instead of `ubuntu-latest`.

* use ubuntu runner for android build